backend/requirements.txt
# Python dependencies
fastapi
sqlalchemy
pyodbc
requests
pytest
pylint
python-dotenv
alembic
ldap3
pyjwt
cryptography
